Add files via upload
D3.js assignment
As mentioned in class the objective of this assignment is to get familiar with the world of web based data visualization using D3.js. As you can see in the advanced examples from the web, the main advantage of working with D3.js are the endless possibilities of personalizing and in particular add animation and interactivity to your visualizations. But in order to get there we have start with the basics and get a good foundation of the general programming methods and visual concepts of d3.js, which is the objective of this assignment. 

I ask you to create 5 different data visualizations based on the content I explained to you in the last lectures. 
Presentation: https://slides.com/sandravizmad/d3-js

Visualizations
Bar chart 
Create one bar chart with either vertical of horizontal bars. The chart should include a scaling function as well as showing the labels of the bars. You can either chose the labels inside of the bars or you use the axis to show the labels. 

Scatter plot 
Create one scatter plot. Apart from the visual encodings of the x and y position I want you to add one more visual encoding radius or the color of the circles (it can be redundant). Please apply to all visual a scaling. Moreover I want you to add the axis to your scatter plot. The way you present the axis in terms of styling is up to you. In the presentation you can find a lot examples representing axis in different ways. 

Interactive chart
I want that you chose either the bar chart or the scatter plot you have created and add to this chart some interactivity, either a click (on a button) or a mouseover option. 

 Animated chart 
I want that you chose either the bar chart or the scatter plot you have created and add some animation to it (transition) The animation can appear on itself or through interactivity (e.g. click on the button) that you created in the other chart. 

Free of choice chart
The final visualization don't need to be useful, it's not about the visualizations it's about the code and creativity. Some of you already mentioned nice ideas in class and I encourage you to get wild here :) Try your ideas! 

Some more examples to fork from 
https://bl.ocks.org/sandravizmad
https://bl.ocks.org/romsson
https://www.d3-graph-gallery.com/index.html
https://bl.ocks.org/nbremer (advanced) 

Editor 
You can chose to either use the blockbuilder environment for coding your visualizations. In this case you can just fork the examples you want to use as a template. The final submission will be the links to your block. Please create one block for each of the 5 visualizations. 
If you prefer working locally you can also write your code in your editor, e.g. ATOM, which I use. In case you use atom this video might be useful for you as it explains some features and typical errors. (One student shared it with me after class :) https://www.youtube.com/watch?v=aiXNKHKWlmY The final submission in this case will be 5 html files and the corresponding datasets if you not saved them online on gist, which I actually recommend. 

Datasets
I want that you use for each of your visualizations a different dataset. 
You can again select a dataset from the following links or choose your own one. https://data.oecd.org/ or https://www.makeovermonday.co.uk/data/ 

Please find below the links to a video on how to save your data online on gist. If you work locally remember to save your dataset in the same folder as your html. 
https://www.youtube.com/watch?v=5c8qQjmc2ic

Style 
Each of your visualization should have a title as I showed you in the html sections. This title as well as any other text that will appear in your visualization should have the same font and it should be different to the one I used. Therefore I ask you chose a new font from the google font webpage and apply this to your visualization. 
Moreover I ask you to think and change the colors and transparencies of the body-background and all the elements that appear in your visualizations like the bars, axis, buttons and any text etc.  The styling part is also very important when creating visualizations, which you can experiment with in this assignment. 

Code
The code should be as clean as possible and each part of the code should include a short comment above (similar to the examples I showed you).

The deadline for this assignment is the 28/11/2020. 
If you have any questions feel free to contact me in slack. 

Enjoy, 

Sandra